First stage of project restructuring done
