+++ .gitignore
@@ -1,4 +1,4 @@
 client/node_modules/
-node_modules/
+server/node_modules/
 .idea/
 .env
