+++ client/src/assets/logo_black.png
