+++ client/src/containers/properties-panel/propertiesPanelApplicationDropdown/PropertiesPanelApplicationDropdown.jsx
