+++ client/src/containers/properties-panel/propertiesView/PropertiesView.css
