+++ client/src/index.js
@@ -2,16 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'antd/dist/antd.css';
 import './index.css';
-import './App.css';
-import App from './App';
+import './containers/App.css';
+import App from './containers/App';
 //import reportWebVitals from './reportWebVitals';
 
-
-ReactDOM.render(
-    <App/>,
-    document.getElementById('root')
-);
-
+ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
