+++ server/package.json
@@ -2,7 +2,7 @@
   "name": "ark-automate",
   "version": "1.0.0",
   "scripts": {
-    "start": "node server",
+    "start": "node server.js",
     "build": "cd client/ && npm install && npm run build"
   },
   "cacheDirectories": [
