+++ server/server.js
@@ -2,7 +2,7 @@ const express = require('express');
 const path = require('path');
 const cluster = require('cluster');
 const numCPUs = require('os').cpus().length;
-const activityDataRetrieval = require('./ActivityDataRetrieval');
+const activityDataRetrieval = require('./services/ActivityDataRetrieval');
 const isDev = process.env.NODE_ENV !== 'production';
 const PORT = process.env.PORT || 5000;
